Reduce database env var duplication, make database env var names more explicit and remove unused TABLE_NAME env var
