+++ src/lambdas/determine_legislation_provisions/index.py
@@ -2,7 +2,6 @@
 
 import datetime
 import logging
-import os
 import urllib.parse
 
 import boto3
@@ -12,23 +11,10 @@ from legislation_provisions_extraction.legislation_provisions import (
     provisions_pipeline,
 )
 from replacer.second_stage_replacer import replace_references_by_paragraph
+from utils.environment_helpers import validate_env_variable
 
 LOGGER = logging.getLogger()
 LOGGER.setLevel(logging.INFO)
-
-
-def validate_env_variable(env_var_name):
-    print(f"Getting the value of the environment variable: {env_var_name}")
-
-    try:
-        env_variable = os.environ[env_var_name]
-    except KeyError:
-        raise Exception(f"Please, set environment variable {env_var_name}")
-
-    if not env_variable:
-        raise Exception(f"Please, provide environment variable {env_var_name}")
-
-    return env_variable
 
 
 def upload_contents(source_key, output_file_content):
