+++ src/lambdas/determine_replacements_abbreviations/index.py
@@ -2,27 +2,14 @@
 
 import json
 import logging
-import os
 
 import boto3
 import spacy
 
+from utils.environment_helpers import validate_env_variable
+
 LOGGER = logging.getLogger()
 LOGGER.setLevel(logging.INFO)
-
-
-def validate_env_variable(env_var_name):
-    LOGGER.debug(f"Getting the value of the environment variable: {env_var_name}")
-
-    try:
-        env_variable = os.environ[env_var_name]
-    except KeyError:
-        raise Exception(f"Please, set environment variable {env_var_name}")
-
-    if not env_variable:
-        raise Exception(f"Please, provide environment variable {env_var_name}")
-
-    return env_variable
 
 
 # isolating processing from event unpacking for portability and testing
