+++ src/lambdas/determine_replacements_legislation/index.py
@@ -1,84 +1,17 @@
 #!/usr/bin/env python3
 
-import base64
 import json
 import logging
-import os
 
 import boto3
 import spacy
 
 from database import db_connection
+from utils.environment_helpers import validate_env_variable
+from utils.initialise_db import init_db_connection
 
 LOGGER = logging.getLogger()
 LOGGER.setLevel(logging.INFO)
-
-
-def validate_env_variable(env_var_name):
-    LOGGER.debug(f"Getting the value of the environment variable: {env_var_name}")
-
-    try:
-        env_variable = os.environ[env_var_name]
-    except KeyError:
-        raise Exception(f"Please, set environment variable {env_var_name}")
-
-    if not env_variable:
-        raise Exception(f"Please, provide environment variable {env_var_name}")
-
-    return env_variable
-
-
-############################################
-# CLASS HELPERS
-############################################
-class getLoginSecrets:
-    def get_secret(self, aws_secret_name, aws_region_name):
-        """
-        Get login secrets for database access
-        """
-        secret_name = aws_secret_name
-        region_name = aws_region_name
-
-        # Create a Secrets Manager client
-        session = boto3.session.Session()
-        client = session.client(service_name="secretsmanager", region_name=region_name)
-
-        try:
-            LOGGER.info(" about to get_secret_value_response")
-            get_secret_value_response = client.get_secret_value(SecretId=secret_name)
-            LOGGER.info("got_secret_value_response")
-
-            # Decrypts secret using the associated KMS CMK.
-            # Depending on whether the secret is a string or binary, one of these fields will be populated.
-            if "SecretString" in get_secret_value_response:
-                secret = get_secret_value_response["SecretString"]
-                LOGGER.info("got SecretString")
-            else:
-                LOGGER.info("not SecretString")
-                decoded_binary_secret = base64.b64decode(
-                    get_secret_value_response["SecretBinary"]
-                )
-                secret = decoded_binary_secret
-            LOGGER.info("here")
-            return secret
-        except Exception as exception:
-            LOGGER.error("Exception: %s", exception)
-            raise
-
-
-############################################
-# - INSTANTIATE CLASS HELPERS
-# - GET ENV VARIABLES
-############################################
-database_name = validate_env_variable("DATABASE_NAME")
-table_name = validate_env_variable("TABLE_NAME")
-username = validate_env_variable("USERNAME")
-port = validate_env_variable("PORT")
-host = validate_env_variable("HOSTNAME")
-aws_secret_name = validate_env_variable("SECRET_PASSWORD_LOOKUP")
-aws_region_name = validate_env_variable("REGION_NAME")
-
-get_secret = getLoginSecrets()
 
 
 # isolating processing from event unpacking for portability and testing
@@ -192,17 +125,6 @@ def init_NLP():
     return nlp
 
 
-def init_DB():
-    """
-    Establish database connection
-    """
-    password = get_secret.get_secret(aws_secret_name, aws_region_name)
-    db_conn = db_connection.create_connection(
-        database_name, username, password, host, port
-    )
-    return db_conn
-
-
 def close_connection(db_conn):
     """
     Close the database connection
@@ -215,7 +137,7 @@ def determine_replacements(file_content):
     Fetch legislation replacements from database
     """
     # connect to the database
-    db_conn = init_DB()
+    db_conn = init_db_connection()
     LOGGER.info("Connected to database")
 
     # setup the spacy pipeline
