+++ src/lambdas/fetch_xml/index.py
@@ -1,27 +1,14 @@
 # Replace this file with functional code rather than one that just lists the S3 buckets.
 import json
 import logging
-import os
 
 import boto3
 import urllib3
 
+from utils.environment_helpers import validate_env_variable
+
 LOGGER = logging.getLogger()
 LOGGER.setLevel(logging.INFO)
-
-
-def validate_env_variable(env_var_name):
-    print(f"Getting the value of the environment variable: {env_var_name}")
-
-    try:
-        env_variable = os.environ[env_var_name]
-    except KeyError:
-        raise Exception(f"Please, set environment variable {env_var_name}")
-
-    if not env_variable:
-        raise Exception(f"Please, provide environment variable {env_var_name}")
-
-    return env_variable
 
 
 ############################################
